requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0
httpx>=0.25.0
websockets>=12.0
pydantic>=2.5.0

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# AI/ML Core
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Diffusion models
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# Face processing
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# Body processing
controlnet-aux>=0.0.7

# Segmentation
segment-anything>=1.0

# Image enhancement
realesrgan>=0.3.0
basicsr>=1.4.2

# Database
aiosqlite>=0.19.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0

# RunPod specific
runpod>=1.3.0
